Make a separate "stats" module, and move the Stats type/implementation/constant/tests there. Tweak the Config#has_stat signature.
