+++ src/output/stdout_table.rs
@@ -1,6 +1,7 @@
 use super::*;
 use crate::data::Data;
-use crate::parse_args::{Config, Stats};
+use crate::parse_args::Config;
+use crate::stats::Stats;
 
 pub fn stdout_table(config: &Config, data: &Data) {
     use humantime::format_duration;
@@ -20,13 +21,13 @@ pub fn stdout_table(config: &Config, data: &Data) {
         Cell::new("Example keys"),
     ]));
 
-    if config.has_stat(Stats::Memory) {
+    if config.has_stat(&Stats::Memory) {
         let row = table.get_mut_row(0).unwrap();
         row.add_cell(Cell::new("Memory (sum)"));
         row.add_cell(Cell::new("Memory (p50/90/99)"));
     }
 
-    if config.has_stat(Stats::TTL) {
+    if config.has_stat(&Stats::TTL) {
         let row = table.get_mut_row(0).unwrap();
         row.add_cell(Cell::new("TTL (% with)"));
         row.add_cell(Cell::new("TTL (p50/90/99)"));
@@ -45,7 +46,7 @@ pub fn stdout_table(config: &Config, data: &Data) {
             Cell::new(&example_keys(&bin).join("\n")),
         ]);
 
-        if config.has_stat(Stats::Memory) {
+        if config.has_stat(&Stats::Memory) {
             row.add_cell(Cell::new(&convert(memory::total(data, bin) as f64)));
             row.add_cell(Cell::new(
                 &vec![
@@ -57,7 +58,7 @@ pub fn stdout_table(config: &Config, data: &Data) {
             ));
         }
 
-        if config.has_stat(Stats::TTL) {
+        if config.has_stat(&Stats::TTL) {
             row.add_cell(Cell::new(&format!("{:.2}%", ttl::pct_with_ttl(data, bin))));
             row.add_cell(Cell::new(
                 &vec![
