+++ src/parse_args.rs
@@ -2,10 +2,9 @@ use clap::crate_version;
 use clap::Clap;
 use glob;
 use std::collections::HashSet;
-use std::str::FromStr;
 
-static STATS_OPTIONS: [&str; 2] = ["memory", "ttl"];
 use crate::output::{OutputMode, OUTPUT_MODE_OPTIONS};
+use crate::stats::{Stats, STATS_OPTIONS};
 
 #[derive(Clap, Eq, PartialEq, Debug)]
 #[clap(version = crate_version!())]
@@ -46,27 +45,8 @@ impl Config {
         self.stats.retain(|s| uniques.insert(s.clone()));
     }
 
-    pub fn has_stat(&self, stat: Stats) -> bool {
-        self.stats.iter().any(|s| *s == stat)
-    }
-}
-
-#[derive(Eq, PartialEq, Hash, Clone, Debug)]
-pub enum Stats {
-    Memory,
-    TTL,
-}
-
-impl FromStr for Stats {
-    type Err = String;
-
-    fn from_str(s: &str) -> Result<Self, Self::Err> {
-        use Stats::*;
-        match s {
-            "memory" => Ok(Memory),
-            "ttl" => Ok(TTL),
-            _ => Err(format!("Unknown value: {}", s)),
-        }
+    pub fn has_stat(&self, stat: &Stats) -> bool {
+        self.stats.iter().any(|s| s == stat)
     }
 }
 
@@ -84,14 +64,6 @@ fn validate_url(url: &str) -> Result<(), String> {
 #[cfg(test)]
 mod tests {
     use super::*;
-
-    #[test]
-    fn config_stats_options() {
-        for opt in &STATS_OPTIONS {
-            opt.parse::<Stats>()
-                .unwrap_or_else(|_| panic!("Unsupported: {}", opt));
-        }
-    }
 
     #[test]
     fn config_validate_url() {
